chore(44545): move extracted file to app/scripts/sidepanel/background.ts
